requirements.txt
ultralytics
gradio
pillow
numpy
